Add support for nested comments
Change the way comments, strings and other sub-sequences are handled:

For IDEA to not get crazy, we have to return a match also for erroneous tokens. E.g. when starting a comment with "/*" the rest of the file should be considered a comment block (Assuming  there is no "*/" somewhere after it). Although this is technically not correct, by lexing the file like this we are done faster. The error can be taken care of in the annotator, which would have to check whether the comment is ended before the file ends.